+++ inst/include/dplyr/visitors/subset/column_subset.h
@@ -15,8 +15,34 @@ SEXP bracket_one();
 SEXP bracket_two();
 }
 
-
 namespace dplyr {
+
+struct SingleIndex {
+  IntegerVector index;
+
+  SingleIndex() :
+    index(IntegerVector::create(1))
+  {}
+
+  inline int size() const {
+    return 1;
+  }
+
+  inline int operator[](int i) const {
+    return index[0] - 1;
+  }
+
+  inline operator SEXP() const {
+    return index;
+  }
+
+  inline SingleIndex& operator++() {
+    ++index[0];
+    return *this;
+  }
+
+};
+
 namespace traits {
 
 template <typename T>
@@ -24,6 +50,10 @@ struct can_mark_na ;
 
 template <>
 struct can_mark_na<IntegerVector> {
+  typedef Rcpp::traits::true_type type;
+};
+template <>
+struct can_mark_na<IntegerVectorView> {
   typedef Rcpp::traits::true_type type;
 };
 template <>
@@ -36,6 +66,10 @@ struct can_mark_na<RowwiseSlicingIndex> {
 };
 template <>
 struct can_mark_na<NaturalSlicingIndex> {
+  typedef Rcpp::traits::false_type type;
+};
+template <>
+struct can_mark_na<SingleIndex> {
   typedef Rcpp::traits::false_type type;
 };